fix: resolve CI failures and complexity violations
- Fix flow order guardrail by updating ALLOWED_VIOLATIONS line number for json_output.py.
- Fix ShadowFork tests by correcting git mock sequences to match implementation.
- Fix Flow Studio UI ID tests by moving dynamic element check to TestDynamicUIIDs.
- Refactor large test files into smaller modules to pass complexity checks:
  - Split `tests/test_flow_order_guardrail.py` -> `tests/test_flow_order_guardrail_config.py`
  - Split `tests/test_flow_studio_ui_ids.py` -> `tests/test_flow_studio_ui_ids_{core,components,integration,dynamic}.py` and helpers
  - Split `tests/test_shadow_fork.py` -> `tests/test_shadow_fork_{core,ops}.py`
